+++ example/lib/main.dart
@@ -4,9 +4,6 @@ import 'package:flame/components/component.dart';
 import 'package:flame/components/mixins/resizable.dart';
 import 'package:flame/gestures.dart';
 import 'package:flame/game.dart';
-import 'package:flame/text_config.dart';
-import 'package:flame/position.dart';
-import 'package:flame/palette.dart';
 import 'package:flame_flare/flame_flare.dart';
 import 'package:flutter/material.dart';
 import 'package:flare_flutter/flare_controls.dart';
@@ -19,8 +16,6 @@ void main() {
 }
 
 class MyGame extends BaseGame with TapDetector, DoubleTapDetector {
-  final TextConfig fpsTextConfig = TextConfig(color: BasicPalette.white.color);
-
   final paint = Paint()..color = const Color(0xFFE5E5E5E5);
   final List<String> _animations = ['Wave', 'Dance'];
   int _currentAnimation = 0;
@@ -64,10 +59,6 @@ class MyGame extends BaseGame with TapDetector, DoubleTapDetector {
   @override
   void render(Canvas canvas) {
     super.render(canvas);
-
-    if (debugMode()) {
-      fpsTextConfig.render(canvas, fps(120).toString(), Position(0, 10));
-    }
   }
 }
 
@@ -93,20 +84,31 @@ class MinionComponent extends FlareActorComponent with Resizable {
   MinionController minionController;
 
   MinionComponent(this.minionController)
-      : super(FlareActorAnimation(
-          'assets/Bob_Minion.flr',
-          controller: minionController,
-          fit: BoxFit.contain,
-          alignment: Alignment.bottomCenter,
-          width: 306,
-          height: 500,
-        ));
+      : super(
+          FlareActorAnimation(
+            'assets/Bob_Minion.flr',
+            controller: minionController,
+            fit: BoxFit.scaleDown,
+            alignment: Alignment.center,
+          ),
+          width: 300,
+          height: 300,
+        );
+
+  @override
+  void render(Canvas c) {
+    final rect = Rect.fromLTWH(x, y, width, height);
+    final paint = Paint()..color = const Color(0xFFfafbfc);
+    c.drawRect(rect, paint);
+    super.render(c);
+  }
 
   @override
   void resize(Size size) {
     super.resize(size);
-    width = size.width;
-    height = size.height;
+
+    x = size.width / 2 - width / 2;
+    y = size.height / 2 - height / 2;
   }
 }