+++ lib/src/flare_animation.dart
@@ -6,8 +6,6 @@ part of flame_flare;
 class FlareActorAnimation {
   FlareActorAnimation(
     this.filename, {
-    @required this.width,
-    @required this.height,
     this.boundsNode,
     this.animation,
     this.fit = BoxFit.contain,
@@ -25,8 +23,6 @@ class FlareActorAnimation {
 
   FlareActorAnimation.asset(
     this.flareProvider, {
-    @required this.width,
-    @required this.height,
     this.boundsNode,
     this.animation,
     this.fit = BoxFit.contain,
@@ -46,9 +42,6 @@ class FlareActorAnimation {
 
   // Flare only allows the render box to be loaded if it is considered "attached", we need this ugly dumb thing here to do that.
   final _pipelineOwner = _FlareActorComponentPipelineOwner();
-
-  /// The animation size base.
-  double width = 0.0, height = 0.0;
 
   // Fields are ported from flare actor widget
   /// Mirror to [FlareActor.filename]
@@ -116,12 +109,13 @@ class FlareActorAnimation {
     _loadRenderBox();
   }
 
-  void render(Canvas canvas, Offset position) {
+  void render(Canvas canvas, ui.Size size) {
     assert(_renderObject != null,
         "FlareActorAnimation was rendered before initialization. Run FlareActorAnimation.init() before rendering it");
     final bounds = _renderObject.aabb;
+
     if (bounds != null) {
-      _paintActor(canvas, bounds, position);
+      _paintActor(canvas, bounds, size);
     }
   }
 
@@ -146,16 +140,21 @@ class FlareActorAnimation {
 
   // Paint procedures ported from FlareRenderBox.paint with some changes that
   // makes sense on a flame context
-  void _paintActor(Canvas c, AABB bounds, Offset position) {
-    final size = ui.Size(width, height);
+  void _paintActor(Canvas c, AABB bounds, ui.Size size) {
+    final position = Offset.zero;
 
     final contentWidth = bounds[2] - bounds[0];
     final contentHeight = bounds[3] - bounds[1];
-    final x = -1 * bounds[0] - contentWidth / 2.0;
-    final y = -1 * bounds[1] - contentHeight / 2.0;
+    final x = -1 * bounds[0] -
+        contentWidth / 2.0 -
+        (alignment.x * contentWidth / 2.0);
+    final y = -1 * bounds[1] -
+        contentHeight / 2.0 -
+        (alignment.y * contentHeight / 2.0);
 
     double scaleX = 1.0, scaleY = 1.0;
 
+    c.save();
     // pre paint
     if (shouldClip) {
       c.clipRect(position & size);
@@ -168,43 +167,49 @@ class FlareActorAnimation {
         scaleY = size.height / contentHeight;
         break;
       case BoxFit.contain:
-        final minScale =
+        double minScale =
             min(size.width / contentWidth, size.height / contentHeight);
         scaleX = scaleY = minScale;
         break;
       case BoxFit.cover:
-        final maxScale =
+        double maxScale =
             max(size.width / contentWidth, size.height / contentHeight);
         scaleX = scaleY = maxScale;
         break;
       case BoxFit.fitHeight:
-        final minScale = size.height / contentHeight;
+        double minScale = size.height / contentHeight;
         scaleX = scaleY = minScale;
         break;
       case BoxFit.fitWidth:
-        final minScale = size.width / contentWidth;
+        double minScale = size.width / contentWidth;
         scaleX = scaleY = minScale;
         break;
       case BoxFit.none:
         scaleX = scaleY = 1.0;
         break;
       case BoxFit.scaleDown:
-        final minScale =
+        double minScale =
             min(size.width / contentWidth, size.height / contentHeight);
         scaleX = scaleY = minScale < 1.0 ? minScale : 1.0;
         break;
     }
 
     final transform = Mat2D();
-    transform[4] = position.dx + size.width / 2.0;
-    transform[5] = position.dy + size.height / 2.0;
+    transform[4] = size.width / 2.0 + (alignment.x * size.width / 2.0);
+    transform[5] = size.height / 2.0 + (alignment.y * size.height / 2.0);
     Mat2D.scale(transform, transform, Vec2D.fromValues(scaleX, scaleY));
     final center = Mat2D();
     center[4] = x;
     center[5] = y;
     Mat2D.multiply(transform, transform, center);
 
+    c.translate(
+      size.width / 2.0 + (alignment.x * size.width / 2.0),
+      size.height / 2.0 + (alignment.y * size.height / 2.0),
+    );
+
     c.scale(scaleX, scaleY);
+    c.translate(x, y);
 
     _renderObject.paintFlare(c, transform);
     c.restore();